% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SeqFasta.R
\name{as.SeqFasta}
\alias{as.SeqFasta}
\title{S3 Object 'SeqFasta'}
\usage{
as.SeqFasta(object, GeneName = NULL, Annotation = NULL)
}
\arguments{
\item{object}{Sequence}
}
\description{
S3 objects designed for Genomic data
}
\examples{
fasta = as.SeqFasta('ATCGATCGATCGATCG',
                    GeneName = 'Potri.000001',
                    Annotation = '>Potri.000001 pacid=1000001 locus=Potri.000001')
fasta
}
