% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SeqFastq.R
\name{as.SeqFastq}
\alias{as.SeqFastq}
\title{S3 Object 'SeqFastq'}
\usage{
as.SeqFastq(object, Qscore = NULL, ReadName = NULL, RunInfo = NULL)
}
\arguments{
\item{object}{Sequence}
}
\description{
S3 objects designed for Genomic data
}
\examples{
fastq = as.SeqFastq('ATCGATCGATCGATCG',
                    Qscore = '#&$(U@(#!*#)!(#*',
                    ReadName = 'ReadName',
                    RunInfo = 'Run_info_from_sequencing_platform')
fastq
}
