% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_template.R
\name{basic_analysis_report_opt}
\alias{basic_analysis_report_opt}
\title{basic analysis report template}
\usage{
basic_analysis_report_opt()
}
\arguments{
\item{toc}{}
}
\value{
call basic analysis report chunk opts
}
\description{
basic analysis report template
}
\examples{
basic_analysis_report_opt()
}
