% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_expression.R
\name{calcZ}
\alias{calcZ}
\title{Transform expression matrix into z score}
\usage{
calcZ(expression_matrix)
}
\arguments{
\item{log2}{expression matrix}
}
\value{
log2 expression z-score
}
\description{
Transform expression matrix into z score
}
\examples{
calcZ(expression_matrix)
}
\keyword{expression}
