% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_expression.R
\name{calcZref}
\alias{calcZref}
\title{Transform expression matrix into z score based on reference expression}
\usage{
calcZref(expression_matrix, reference_matrix)
}
\arguments{
\item{log2}{expression matrix, log2 reference expression matrix}
}
\value{
log2 expression z-score
}
\description{
Transform expression matrix into z score based on reference expression
}
\examples{
calcZref(expression_matrix, reference_matrix)
}
\keyword{expression}

