% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkPhenotype.R
\name{checkPhenotype}
\alias{checkPhenotype}
\title{Check the phenotype data frame}
\usage{
checkPhenotype(data, formula = NULL, id = NULL, gender = NULL,
  include = NULL, exclude = NULL)
}
\arguments{
\item{data}{an object of type \code{phenotype} or a data frame (or object
coercible by as.data.frame to a data frame) containing the variables in the
model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.The details
of model specification are given under 'Details' in the "lm" help file.}

\item{id}{(optional) column name identifier of the unique subjects in
\code{data}.  If given the phenotype sample ids will become the rownames
of \code{data}. If \code{id} is \code{NULL} the suject id's are assumed
to be the rownames of \code{data}.}

\item{gender}{(optional) column name identifier for the gender in \code{data}}

\item{include}{(optional) character vector of the subjects in \code{id} to be
included in the analysis.  See Details.}

\item{exclude}{(optional) character vector of the subjects in \code{id} to be
excluded in the analysis.  See Details.}
}
\value{
invisible(NULL)
}
\description{
This function provides a basic check that the phenotype file is
  a suitable for further analysis.  It checks that the required column names
  are in the phenotype data frame and that subjects are unique.
}
\details{
TBD
}

