% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{classHeatmap}
\alias{classHeatmap}
\title{classHeatmap}
\arguments{
\item{data}{A data frame with columns \emph{gene} and \emph{class} indicating
the class specific gene expression.}

\item{counts.log}{A matrix of log normalized counts.}

\item{classes}{A tibble with columns \emph{sample} and \emph{class}
indicating the sample ID and class (cell type) of the sample.}
}
\description{
classHeatmap
}
\author{
Jason T. Serviss
}
\keyword{classHeatmap}
