% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_probes_2_gene.R
\name{combine_probes_2_gene}
\alias{combine_probes_2_gene}
\title{Combine multiple probes to gene level expression using svd}
\usage{
combine_probes_2_gene(expr, genes, method = "svd")
}
\arguments{
\item{genes}{from rownames of expression data frame}

\item{method}{(svd)}

\item{expression}{dataframe}
}
\description{
Combine multiple probes to gene level expression using svd
}
\examples{
combine_probes_2_genes(exprs_df, genes)
}
