% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGene.R
\name{convertGeneSymbols}
\alias{convertGeneSymbols}
\title{Convert gene names}
\usage{
convertGeneSymbols(genes, species1, species2)
}
\arguments{
\item{a}{list of genes symbols}

\item{species}{from}

\item{species}{to}
}
\value{
a data frame of genes symbol lookup
}
\description{
convert gene symbol between hg19, mm10, rn6
}
\examples{
genes_df = convertGeneSymbols(genes, "mm10", "hg19")
}
\keyword{genes}

