% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{DEG_analysis}
\alias{DEG_analysis}
\title{Analyze Differentially Expressed Genes From RNA-seq Data}
\usage{
DEG_analysis(
  countData,
  group,
  contrast = NULL,
  p.value = 1,
  logFC = 0,
  method = "DESeq2",
  FilterGene = T,
  log_transformed = NULL
)
}
\arguments{
\item{countData}{A standard expression matrix with rows the genes and columns the samples. The matrix can be raw count or normalized according to the method chose.}

\item{group}{A vector of groups of the sample. Note that only 2 groups are supported.}

\item{contrast}{A string with the format "A-B", in which "A" and "B" are to groups in the group parameter.
The result demonstrates the gene expression in group A vs. that in group B if input is "A-B".
If NULL, the result will denote unique(group)[1] vs. unique(group)[2].}

\item{p.value}{A number denotes the largest adjusted p value in the result table.}

\item{logFC}{A number denotes the minimum absolute log2 fold change in the result table.}

\item{method}{A string tells the method used for DEG analysis, can be "DESeq2", "edgeR", "limma", "wilcox" or "t".
DESeq2 uses a negative binomial model. It accepts raw count data and is the default method.
edgeR uses a negative binomial model as well but uses a different normalization method. It accepts raw count data.
limma uses a linear model and uses Bayesian methods to reduce false positive results. It accepts log2-transformed raw count or TPM data.
wilcox performs Wilcoxn test to every gene. Suitable for data with a large sample size. No require on data distribution.
t performs t test to every gene. Suitable for data following a normal distribution and variances are homogeneous. Not recommended for most cases.}

\item{FilterGene}{A logical value denotes whether filter genes with low expression to reduce false positive rate. Only works for limma, wilcox and t methods.}

\item{log_transformed}{A logical value denotes whether the input is log-transformed. Default is NULL and the function detects it automatically. Only when the detection fails should it be specified.}
}
\value{
A result table containing log2 fold change, p value and p value adjusted by FDR method.
}
\description{
Get DE genes using one of several common methods.
}
\examples{
res <- DEG_analysis(test, group = c("A", "A", "A", "B", "B", "B"), contrast = "B-A", method = "wilcox", Normalized = T, FilterGene = F)
}
