% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw heatmap}
\usage{
draw_heatmap(data, group = NULL, scale = TRUE, cutoff = 0.1, ...)
}
\arguments{
\item{data}{A data.frame/matrix with row the variables and the column samples.}

\item{group}{A vector specifies the group of the samples in data.}

\item{scale}{Whether to perform Z-score scale.}

\item{cutoff}{The cutoff for color shown. e.g., 0.1 means the lowest 10 percent/highest 10 percent will be shown as the darkest color.}

\item{...}{Other parameters for ComplexHeatmap::Heatmap}
}
\value{
A Heatmap object.
}
\description{
A quick tool for drawing heatmap.
}
