% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{draw_pca}
\alias{draw_pca}
\title{Draw pca plot}
\usage{
draw_pca(
  data,
  group = NULL,
  ncp = 2,
  draw = T,
  label = "none",
  addEllipses = T,
  palette = c("#00AFBB", "#E7B800", "#FC4E07"),
  geom.ind = c("point")
)
}
\arguments{
\item{data}{A data.frame with row the variables and the column samples.}

\item{group}{A vector specifies the group of the samples in data.}

\item{ncp}{A number specifies the principal components in the result. Default is 2.}

\item{label, mean.point, addEllipses, palette}{refer to factoextra::fviz_pca_ind}
}
\description{
Draw pca plot
}
