% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{draw_ratio}
\alias{draw_ratio}
\title{Draw ratio of different cell populations.}
\usage{
draw_ratio(
  anno.by,
  group.by,
  y.axis = "Fraction of clusters in each group",
  width = 0.5,
  return_table = F
)
}
\arguments{
\item{group.by}{A vector from meta.data showing the condition of cells.}

\item{y.axis}{The title on the y.axis.}

\item{width}{width of columns in the figure.}

\item{return_table}{If TRUE, return a table of cell proportions. If FALSE, return a figure showing the ratio.}

\item{Anno.by}{A vector from meta.data showing the annotation of cells.}
}
\description{
Draw ratio of different cell populations.
}
\examples{
p <- draw_ratio(scRNA$cell_type, scRNA$treatment)
}
