% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{draw_Venn}
\alias{draw_Venn}
\title{Draw Venn plot}
\usage{
draw_Venn(venn_list, color = NULL, ...)
}
\arguments{
\item{venn_list}{A list of vectors for overlapping.}

\item{color}{A list of colors for different groups}

\item{...}{Other parameters for ggvenn::ggvenn}
}
\description{
Draw Venn plot
}
