% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{draw_Volcano}
\alias{draw_Volcano}
\title{Draw Volcano plot for RNAseq result}
\usage{
draw_Volcano(
  deg,
  p.value = 0.05,
  logFC = 1,
  color = c("#6a82ed", "grey", "#ed6f6f"),
  alpha = 0.3,
  size = 2,
  label = NULL,
  highlight = NULL,
  highlight_color = "#07a818",
  max.overlaps = 10
)
}
\arguments{
\item{deg}{Result of DEG_analysis(), or a data.frame with at least two columns "log2FoldChange" and "padj".}

\item{p.value}{The p.value cutoff for identifying the significant DEGs.}

\item{logFC}{The logFC cutoff for identifying the significant DEGs.}

\item{color}{A vector of three colors denoting the down regulated, not significant, and up regulated dots.}

\item{label}{A vector of genes that require to show their names.}

\item{highlight}{A vector of genes that require to highlight their dots.}

\item{highlight_color}{Specify the color of highlight.}

\item{max.overlaps}{refer to ggrepel::geom_label_repel()}
}
\description{
Draw Volcano plot for RNAseq result
}
