% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentScore.R
\name{enrichmentScore}
\alias{enrichmentScore}
\title{enrichmentScore}
\usage{
enrichmentScore(cpm, classes, e1 = 0.1, e2 = 0.01)
}
\arguments{
\item{cpm}{matrix; Matrix containing cpm values.}

\item{classes}{Character; Classes corresponding to columns in cpm argument.}
}
\value{
Returns the a list of enrichment scores with one element per
classification.
}
\description{
Calculates an the enrichment of genes per classification.
https://www.biorxiv.org/content/biorxiv/early/2018/04/06/294918.full.pdf
}
\author{
Jason T. Serviss
}
\keyword{nTopVar}
