% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{Extract_Expression_Matrix}
\alias{Extract_Expression_Matrix}
\title{Extract expression matrix from Seurat Object}
\usage{
Extract_Expression_Matrix(scRNA, Normalized = F)
}
\arguments{
\item{Normalized}{}
}
\description{
Extract expression matrix from Seurat Object
}
