% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{Find_topn_markers}
\alias{Find_topn_markers}
\title{Find markers for each cell cluster}
\usage{
Find_topn_markers(scRNA, n = 10, draw = T, high_expr = T, SCTransform = F, ...)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{n}{Top n markers.}

\item{draw}{If TRUE, return a figure showing the markers. If FALSE, return the marker genes.}

\item{high_expr}{Whether the markers genes must have a high log fold change compared to other clusters.}

\item{SCTransform}{Whether the object is SCTransformed.}

\item{...}{Refer to FindAllMarkers()}
}
\description{
Find markers for each cell cluster
}
