% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{findHighlyCorrelated}
\alias{findHighlyCorrelated}
\title{findHighlyCorrelated}
\arguments{
\item{cpm}{Matrix; Matrix containing cpm values.}

\item{cut}{Integer; Genes with a correlation > cut will be reported.}
}
\description{
Identifies genes with a level of correlation greater than a specified cutoff.
Returns indexes of all genes that are NOT correlated > the specified cutoff.
}
\author{
Jason T. Serviss
}
\keyword{recursiveFE}
