% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{foldChangePerClass}
\alias{foldChangePerClass}
\title{foldChangePerClass}
\arguments{
\item{counts}{The matrix holding expression values.}

\item{class}{A tibble with columns \emph{class} and \emph{sample} indicating
the class and sample ID respectivley.}
}
\description{
Calculates fold change for each gene with each class vs all other classes.
}
\author{
Jason T. Serviss
}
\keyword{foldChangePerClass}
