% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{get_standard_gene_length}
\alias{get_standard_gene_length}
\title{To get gene.length from hg39 when the RNAseq length info is missing (not recommended)}
\usage{
get_standard_gene_length(countData, species, input_datatype)
}
\description{
To get gene.length from hg39 when the RNAseq length info is missing (not recommended)
}
