% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNameField.R
\name{getNameField}
\alias{getNameField}
\alias{getNameField.SeqFasta}
\alias{getNameField.SeqFastq}
\title{Get Namefield From Read Annotation}
\usage{
getNameField(object, split = NULL)

\method{getNameField}{SeqFasta}(object, split = "=| |[|]")

\method{getNameField}{SeqFastq}(object, split = "=| |[|]")
}
\arguments{
\item{object}{class 'SeqFastq'/'SeqFasta' object}

\item{split}{characters to split the annotation line}
}
\description{
This function extract keywords from read annotation.
}
\examples{
file = system.file('data/ref.fa', package = 'seqtools')
fasta = read.fasta(file)
getNameField(fasta[[1]])
}
