% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(name)
}
\arguments{
\item{name}{the name of the person to say hello to}
}
\value{
a printed message
}
\description{
hello function
}
\examples{

hello("John")

}
