% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_Enricher.R
\name{kegg_metabolite_enrichment}
\alias{kegg_metabolite_enrichment}
\title{Perform KEGG metabolite enrichment}
\usage{
kegg_metabolite_enrichment(
  KEGGid,
  species,
  p.adjust.method = "BH",
  population.id = NULL
)
}
\arguments{
\item{KEGGid}{A character vector of KEGG compound IDs (e.g., "C00160")}

\item{species}{A character string, KEGG species code (e.g., "hsa", "eco")}

\item{p.adjust.method}{Adjustment method for multiple testing, default is "BH"}

\item{population.id}{An integer denoting the number of all the compounds that is mapped by KEGG, used for targeted metabolomics}
}
\value{
A data.frame with enrichment results
}
\description{
Perform KEGG metabolite enrichment
}
