% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rsem.R
\name{load_rsem}
\alias{load_rsem}
\title{Load RSEM gene expression}
\usage{
load_rsem(rsem_path, count_type = NULL)
}
\arguments{
\item{path}{to rsem data}

\item{count}{type: "expected_count"(default), "tmp", "fpkm"}
}
\value{
a data.frame of rsem expected counts (no transformations whatsoever). rownames are genes, colnames are samples
}
\description{
Load RSEM gene expression
}
\examples{
exprs = load_rsem("/path/rsem/")
}
\keyword{rsem}
