% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{normalization}
\alias{normalization}
\title{Perform Normalization to RNAseq data}
\usage{
normalization(countData, method, group = NULL, gene.length = NULL)
}
\arguments{
\item{countData}{A standard count matrix with rows the genes and columns the samples.}

\item{method}{Can be "cpm", "rpkm", "tpm", "tmm" or "vst/rlog".}

\item{group}{Used for vst and tmm method.}

\item{gene.length}{The length of genes that required to be normalized. If NULL, use internal data to get gene length.}

\item{species}{"Hs" for human and "Mm" for mouse. Only works when gene.length = NULL.}

\item{input_datatype}{The gene format of the input data. Only works when gene.length = NULL.}
}
\description{
Perform Normalization to RNAseq data
}
