% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_probes_2_gene.R
\name{oneToManyOperation}
\alias{oneToManyOperation}
\title{one to many operations helper for combine_probes_2_genes (local)}
\usage{
oneToManyOperation(one, many, func = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
one to many operations helper for combine_probes_2_genes (local)
}
