% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{predict_SingleR}
\alias{predict_SingleR}
\title{Quick annotate cell clusters using SingleR}
\usage{
predict_SingleR(scRNA, ref, labels, return_labeled = F)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{ref}{Reference data.}

\item{labels}{Choose the labels in the reference data.}

\item{return_labeled}{If TRUE, return a Seurat object that is already annotated. If FALSE, return a table containing the annotation.}
}
\description{
Quick annotate cell clusters using SingleR
}
