% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTest.R
\name{processKStest}
\alias{processKStest}
\title{processKStest}
\arguments{
\item{results}{Write something}

\item{classes}{Write something}

\item{alpha}{Write something}
}
\value{
Write something
}
\description{
Identifies genes for classes that were significant in all comparisons using
the KStest function. In addition, calculates the sum of statistics for each
gene and cell type.
}
\examples{
#nothing here yet
}
\author{
Jason T. Serviss
}
