% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{pseudoBulk}
\alias{pseudoBulk}
\title{Running pseudobulk to find DEGs in scRNA-seq}
\usage{
pseudoBulk(
  scRNA,
  idents_to_check,
  group = NULL,
  replicates = NULL,
  aggregate_expr = T,
  p.value = 1,
  min.pct = 0,
  logfc = 0,
  test.use = "DESeq2"
)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{idents_to_check}{The target idents to be checked. Name is decided by group.
For example, idents are grouped by "feature1" (including "A" and "B") and "feature2" (including "a" and "b"),
then the idents are "A_a", "A_b", "B_a" and "B_b".}

\item{group}{A character vector specifying how to group cells. Elements must exist in metadata.}

\item{replicates}{A character vector specifying for each group, what are the replicates. Note that replicates should not be specified in "idents_to_check"}

\item{aggregate_expr}{Whether to aggregate the expression matrix as a pseudobulk matrix. TRUE is recommended, especially when replicates exist.}

\item{p.value}{The p.value cutoff for result. Default is 1.}

\item{min.pct}{The minimum percent of cells that express one gene. Default is 0.}

\item{logfc}{The logFC cutoff for result. Default is 0.}

\item{test.use}{Refer to FindMarkers().}
}
\description{
Running pseudobulk to find DEGs in scRNA-seq
}
