% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_expression.R
\name{quantnorm}
\alias{quantnorm}
\title{Quantile normalize function}
\usage{
quantnorm(expression_matrix)
}
\arguments{
\item{expression}{matrix}
}
\value{
quantile normalized expression data.frame
}
\description{
Quantile normailze expression matrix
}
\examples{
quantnorm(expression_matrix)
}
\keyword{expression}

