% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_expression.R
\name{quantnormref}
\alias{quantnormref}
\title{Quantile normalize function using reference expression matrix}
\usage{
quantnormref(expression_matrix, reference_matrix)
}
\arguments{
\item{expression}{matrix, reference expression matrix}
}
\value{
quantile normalized expression data.frame
}
\description{
Quantile normailze expression matrix with reference expression matrix
}
\examples{
quantnormref(expression_matrix, reference_matrix)
}
\keyword{expression}
