% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{quick_data_reader}
\alias{quick_data_reader}
\title{Quick Reader}
\usage{
quick_data_reader(
  dir_path,
  sep,
  pattern_index,
  min.cells = 3,
  min.features = 200
)
}
\arguments{
\item{dir_path}{A file path denoting the position of single cell raw data, including files like "matrix", "feature", "barcode".}

\item{sep}{The symbol separating the files, e.g., in "GSM123456_Sample1_barcode", the sep is "_"}

\item{pattern_index}{The index of character that denotes the sample name after separation. e.g., in "GSM123456_Sample1_barcode", index is 2}
}
\description{
Quick Reader
}
