% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{quick_enrich}
\alias{quick_enrich}
\title{Perform quick enrich analysis}
\usage{
quick_enrich(
  geneList,
  genesets = "GO",
  species,
  input_datatype,
  universe = NULL,
  p.value = 0.05,
  minGSSize = 10,
  maxGSSize = 500
)
}
\arguments{
\item{geneList}{A geneList including names of significant differentially expressed genes}

\item{genesets}{The geneset for enrich analysis. Can be "GO", "KEGG", and "Reactome"}

\item{species}{"Hs" for human and "Mm" for mouse.}

\item{input_datatype}{The gene format of the input data.}

\item{universe}{The background geneset for ORA analysis.}

\item{minGSSize}{The minimum geneset size for analysis.}

\item{maxGSSize}{The maximum geneset size for analysis.}

\item{q.value}{The cut off adjusted p for enrich analysis.}
}
\description{
Perform quick enrich analysis
}
