% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{quick_GSEA}
\alias{quick_GSEA}
\title{Perform quick GSEA analysis}
\usage{
quick_GSEA(
  geneList,
  genesets = "GO",
  species,
  input_datatype,
  p.value = 0.05,
  minGSSize = 10,
  maxGSSize = 500
)
}
\arguments{
\item{geneList}{A named geneList including the ranking metrics of all genes with names the gene name}

\item{genesets}{The geneset for GSEA analysis. Can be common genesets including "GO", "KEGG", "Reactome", or custom gmtfile.}

\item{species}{"Hs" for human and "Mm" for mouse.}

\item{input_datatype}{The gene format of the input data.}

\item{p.value}{The cut off of GSEA analysis.}

\item{minGSSize}{The minimum geneset size for analysis.}

\item{maxGSSize}{The maximum geneset size for analysis.}
}
\description{
Perform quick GSEA analysis
}
