% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{Quick_ID_conversion}
\alias{Quick_ID_conversion}
\title{Convert Gene ID Conveniently}
\usage{
Quick_ID_conversion(
  geneList,
  species,
  input_datatype,
  target_datatype,
  matrix = F,
  single_mapping = T,
  to_rownames = F
)
}
\arguments{
\item{geneList}{A vector of gene ids or a matrix contains a column of gene ids named "Geneid".}

\item{species}{A string denotes the species. "Hs" for human and "Mm" for mouse.}

\item{input_datatype}{A string denotes the id type in the geneList.}

\item{target_datatype}{A string denotes the target id type.}

\item{matrix}{A logical value denotes whether the geneList is a matrix. TRUE if geneList is a matrix.}

\item{single_mapping}{A logical value denotes whether one input gene id is mapped to only one target gene id, even though it has several alias. Default to be true.}

\item{to_rownames}{For matrix = TRUE only. A logical value denotes whether the converted ids are rownames of the resultant matrix. If rownames are duplicated, the resultant row takes the mean value.}
}
\value{
A vector or matrix.
}
\description{
Convert Gene ID Conveniently
}
\examples{
SYMBOL <- Quick_ID_conversion(ENTREZID, species = "Mm", input_datatype = "ENTREZID", target_datatype = "SYMBOL", matrix = F, single_mapping = T, to_rownames = F)
}
