% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{quick_manual_annotation}
\alias{quick_manual_annotation}
\title{Quick annotate cell clusters.}
\usage{
quick_manual_annotation(scRNA, celltype, pos)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{celltype}{The specific celltype.}

\item{pos}{The number of target seurat cluster (start from 0).}
}
\value{
A Seurat object with metadata having a column "cell_type".
}
\description{
Quick annotate cell clusters.
}
