% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{quick_single_cell}
\alias{quick_single_cell}
\title{A simple pipeline for quick single cell analysis}
\usage{
quick_single_cell(
  dataList,
  species,
  SCTransform = F,
  DoubletFind = T,
  filter = c(300, 5000, 15),
  progress_saving = F,
  CellCycleScoring = F,
  vars.to.regress = c("percent.mt"),
  PCs = NULL,
  resolution = seq(0.2, 1, 0.2),
  samples = NULL
)
}
\arguments{
\item{dataList}{Refer to corresponding functions}

\item{species}{}

\item{resolution}{}
}
\description{
A simple pipeline for quick single cell analysis
}
\examples{
scRNA <- quick_single_cell(dataList = dataList, species = "Hs", SCTransform = T)
}
