% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.R
\name{read.fasta}
\alias{read.fasta}
\title{Read File in Fasta format}
\usage{
read.fasta(fname)
}
\arguments{
\item{fname}{filename}
}
\description{
Read fasta, return a list of 'SeqFasta' object
}
\examples{
file = system.file('data/ref.fa', package = 'seqtools')
fasta = read.fasta(file)
head(fasta)

}
