% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fastq.R
\name{read.fastq}
\alias{read.fastq}
\title{Read File in Fastq format}
\usage{
read.fastq(fname)
}
\arguments{
\item{fname}{filename}
}
\description{
Read fasta, return a list of 'SeqFastq' object
}
\examples{
file = system.file('data/Long-read.fastq', package = 'seqtools')
fastq = read.fastq(file)
head(fastq)

}
