% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{recursiveFE}
\alias{recursiveFE}
\title{recursiveFE}
\arguments{
\item{cpm}{Matrix; Matrix containing cpm values.}

\item{classes}{Character; classifications corresponding to the cpm columns.}

\item{nCV}{Integer; Number of cross-fold validations.}

\item{s}{Numeric; Vector indicating the number of features to select in
each testing round.}
}
\description{
Recursive Feature Elimination; First, the algorithm fits the model to all
predictors. Each predictor is ranked using it’s importance to the model. Let
S be a sequence of ordered numbers which are candidate values for the number
of predictors to retain (S1 > S2, …). At each iteration of feature selection,
the Si top ranked predictors are retained, the model is refit and performance
is assessed. The value of Si with the best performance is determined and the
top Si predictors are used to fit the final model.
}
\author{
Jason T. Serviss
}
\keyword{recursiveFE}
