% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reducePhenotype.R
\name{reducePhenotype}
\alias{reducePhenotype}
\title{Reduces the phenotype data frame}
\usage{
reducePhenotype(p)
}
\arguments{
\item{p}{an object of class 'phenotype'.}
}
\value{
an object of class 'phenotype'.  \code{data} is minimal, \code{include} is set to, \code{exclude} is set to
}
\description{
This function reduces a data set to only the variables used in
a model removing subjects with missing data.
}
\details{
This function reduces a data set to only the variables used in a
  model removing subjects with missing data.  Also, it makes the row names of
  the resulting data fram the subject identifier.  If a gender column is
  specified the column identifiying the gender will be retained.
}

