% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{rescaleFeatures}
\alias{rescaleFeatures}
\title{Rescale features}
\arguments{
\item{counts}{matrix; a numeric matrix of counts.}
}
\value{
The rescaled matrix.
}
\description{
Rescales each matrix row to the interval \link{0, 1}.
}
\author{
Jason T. Serviss
}
