% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlen.R
\name{rlen}
\alias{rlen}
\title{Calculate Read Length}
\usage{
rlen(object)
}
\arguments{
\item{object}{class 'SeqFasta' or 'SeqFastq' object}
}
\description{
calculate read length
}
\examples{
file = system.file('seqtools/data/long-read.fastq')
fastq = read.fastq(file)
read = fastq[[1]]
rlen(read)

}
