% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription_Seq.R
\name{s2nCalulator}
\alias{s2nCalulator}
\title{Get Signal-to-Noise ranking metrics for GSEA}
\usage{
s2nCalulator(tpm, group, contrast = NULL, abs = F)
}
\arguments{
\item{tpm}{A tpm-normalized expression matrix}

\item{group}{A vector of groups of the sample. Note that only 2 groups are supported.}

\item{contrast}{A string with the format "A-B", in which "A" and "B" are to groups in the group parameter.
The result demonstrates the gene expression in group A vs. that in group B if input is "A-B".
If NULL, the result will denote unique(group)[1] vs. unique(group)[2].}

\item{abs}{Whether to get the absolute values.}
}
\value{
A named, decreasingly ordered numeric vector.
}
\description{
Get Signal-to-Noise ranking metrics for GSEA
}
