% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{sample_object}
\alias{sample_object}
\title{Sample Seurat Object}
\usage{
sample_object(scRNA, proportion)
}
\arguments{
\item{scRNA}{A Seurat Object of a list of Seurat Objects}

\item{proportion}{The proportion of cells in the sample}
}
\description{
Sample Seurat Object
}
