% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{sc_BoxPlot}
\alias{sc_BoxPlot}
\title{Single Cell Draw Boxplot}
\usage{
sc_BoxPlot(scRNA, genes_to_check, group.by)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{group.by}{A character denoting the group.}

\item{n}{Genes to check.}
}
\description{
Single Cell Draw Boxplot
}
