% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{sc_cor}
\alias{sc_cor}
\title{Get correlation in single cell data}
\usage{
sc_cor(
  scRNA,
  genes.check,
  group.by,
  ident.group,
  target.group,
  k = 25,
  max_shared = 10,
  min.cells = 100,
  name = "test",
  method = "pearson"
)
}
\description{
Get correlation in single cell data
}
