% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{score_geneset_activity}
\alias{score_geneset_activity}
\title{Score the specified genesets in Single cell object}
\usage{
score_geneset_activity(scRNA, geneSets, method = "AUCell")
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{geneSets}{A list of genesets or a whole geneset dataframe categorized by column "term"}

\item{method}{A character either "AUCell" or "ModuleScore". "AUCell" is an extra R package and it is better to set seeds because it use random numbers.
"ModuleScore" use internal Seurat methods.}
}
\value{
A Seurat object with activity in meta.data.
}
\description{
Score the specified genesets in Single cell object
}
