% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{scRNA_clustering}
\alias{scRNA_clustering}
\title{Clustering cells}
\usage{
scRNA_clustering(
  scRNA,
  PCs = NULL,
  resolution = seq(0.2, 1, 0.2),
  reduction = "harmony",
  SCTransform = F,
  n.neighbors = 30L,
  progress_saving = F
)
}
\arguments{
\item{scRNA}{Result of scRNA_Integration or single Seurat object after normalization.}

\item{PCs}{Dims used for clustering. If NULL, automatically choose the dim that can explain 90 percent variance.}

\item{resolution}{Resolution for clustering. Higher resolution means more communities. It can be a numeric vector, and if so, the function will ask the optimal choice.}

\item{reduction}{"harmony" if Integration is performed, otherwise "pca" can be used.}

\item{SCTransform}{Whether SCTransform is performed.}

\item{progress_saving}{Whether to save the object after running the function.}
}
\value{
A Seurat object with cells clustered.
}
\description{
Clustering cells
}
\examples{
scRNA <- scRNA_clustering(scRNA, PCs = 25, resolution = seq(0.1, 1.5, 0.1))
}
