% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{scRNA_Integration}
\alias{scRNA_Integration}
\title{Integrate single-cell data}
\usage{
scRNA_Integration(
  Merge.Seurat,
  SCTransform = F,
  theta = NULL,
  progress_saving = F
)
}
\arguments{
\item{Merge.Seurat}{The result returned by scRNA_Normalization_Reduction() or scRNA_SCTransform(). Must have multiple layers.}

\item{SCTransform}{Whether SCTransform is performed.}

\item{progress_saving}{Whether to save the object after running the function.}
}
\value{
A Seurat Object which has cells mapped together for downstream visualization.
}
\description{
Integrate single-cell data
}
\examples{
scRNA <- scRNA_Integration(scRNA)
}
