% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{scRNA_Normalization_Reduction}
\alias{scRNA_Normalization_Reduction}
\title{Perform standard normalization and dimension reduction for Seurat object}
\usage{
scRNA_Normalization_Reduction(
  dataList,
  species,
  CellCycleScoring = F,
  vars.to.regress = c("percent.mt"),
  progress_saving = F
)
}
\arguments{
\item{dataList}{The result returned by scRNA_preprocessing()}

\item{species}{"Hs" for human and "Mm" for mouse, used for Cell Cycle Scoring.}

\item{CellCycleScoring}{Whether to regress the factor of cell cycle gene expression.}

\item{vars.to.regress}{Factors expected to be regressed. If CellCycleScoring = T,"S.Score", "G2M.Score" will be automatically added.}

\item{progress_saving}{Whether to save the object after running the function.}
}
\value{
A merged Seurat Object which has performed normalization, finding variable features, scale and PCA.
}
\description{
Perform standard normalization and dimension reduction for Seurat object
}
\examples{
scRNA <- scRNA_Normalization_Reduction(dataList = dataList, species = "Hs")
}
