% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{scRNA_preprocessing}
\alias{scRNA_preprocessing}
\title{Preprocessing of single-cell data}
\usage{
scRNA_preprocessing(
  dataList,
  species,
  DoubletFind = T,
  filter = c(300, 5000, 15),
  progress_saving = F,
  samples = NULL
)
}
\arguments{
\item{dataList}{Must be a list containing Seurat objects.}

\item{species}{"Hs" for human and "Mm" for mouse, used for estimating mitochondria DNA percent.}

\item{DoubletFind}{Whether to remove doublets.}

\item{filter}{A numeric vector containing filtering criteria for min features, max features, and mitochondria DNA percent.}

\item{progress_saving}{Whether to save the object after running the function.}

\item{samples}{A string of the name in meta.data for grouping cells, only works when DoubletFind = T.}
}
\value{
A dataList with Seurat objects preprocessed.
}
\description{
Preprocessing of single-cell data
}
\examples{
dataaList <- scRNA_preprocessing(dataList = dataList, species = "Hs", filter = c(200,10000,15))
}
