% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell.R
\name{scRNA_SCTransform}
\alias{scRNA_SCTransform}
\title{Perform SCTransform for Seurat object}
\usage{
scRNA_SCTransform(
  dataList,
  vars.to.regress = c("percent.mt"),
  progress_saving = F
)
}
\arguments{
\item{dataList}{The result returned by scRNA_preprocessing()}

\item{vars.to.regress}{Factors expected to be regressed.}

\item{progress_saving}{Whether to save the object after running the function.}
}
\value{
A merged Seurat Object which has performed SCTransform. Not compatible with scRNA_Normalization_Reduction()
}
\description{
Perform SCTransform for Seurat object
}
\examples{
scRNA <- scRNA_SCTransform(dataList = dataList)
}
