% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNameField.R
\name{setNameField}
\alias{setNameField}
\alias{setNameField.SeqFasta}
\alias{setNameField.SeqFastq}
\title{Set Namefield of a Read}
\usage{
setNameField(object, method = NULL, keyValue = NULL, sep = NULL)

\method{setNameField}{SeqFasta}(object, method = "e", keyValue = NULL, sep = " ")

\method{setNameField}{SeqFastq}(object, method = "e", keyValue = NULL, sep = " ")
}
\arguments{
\item{object}{class 'SeqFastq'/'SeqFasta' object}

\item{method}{'r': rewrite, 'e': extend}

\item{keyValue}{list(key = value)}
}
\description{
This function extend/rewrite read
}
\examples{
file = system.file('data/ref.fa', package = 'seqtools')
fasta = read.fasta(file)
getNameField(fasta[[1]])
fasta[[1]] = setNameField(fasta[[1]], method = 'e', keyValue = list(newkey = 'value'))
fasta[[1]]

}
