% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_phenotype.R
\name{summary.phenotype}
\alias{summary.phenotype}
\title{Summarize the phenotype data frame}
\usage{
\method{summary}{phenotype}(p, max.levels = 5L)
}
\arguments{
\item{p}{an object of class 'phenotype'.}

\item{max.levels}{maximum number of \code{levels}.  See Details}
}
\value{
an object of class summary_phenotype.

For numeric variables
\itemize{
 \item \code{N} {the number of non-missing data points in the sample.}
 \item \code{Nmiss} {the number of missing data points in the sample.}
 \item \code{Min} {the sample minima.}
 \item \code{Max} {the sameple maxima.}
 \item \code{Median} {the sample median.}
 \item \code{Mean} {the sample mean.}
 \item \code{SD} {the sample standard deviation.}
 \item \code{Q1} {the sample first quantile.}
 \item \code{Q3} {the sample third quantile.}
}

For (potientially) categorical varaibles
\itemize{
 \item \code{N} {the number of non-missing data points in the sample.}
 \item \code{Nmiss} {the number of missing data points in the sample.}
 \item \code{Levels} {the number of unique values in the sample.}
 \item \code{Datatype} {the (R internal) type or storage mode of the sample.}
}

For (potientially) categorical varaibles with \code{Levels <= max.levels}  a
contingency table of counts for each level.
}
\description{
This function summarizes the phenotype data in a phenotype object.
}
\details{
This is an S3 Methdod that calculates basic summary statistics for phenotype data.

For varaibles taking discrete values (\code{logical}, \code{integer},
\code{charater}, and \code{factor}) the number of unique values, including
\code{NA}, are determined.  If this is less than or equal to max.levels, then
a \code{table} of that varaible is calcualted.  Note: This has no bearing on
whether a variable is treated as a categorical varaible in subsequent
analysis.
}

