% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxocomm.R
\name{taxocomm}
\alias{taxocomm}
\title{Create taxonomy-aggregated community data.frame}
\usage{
taxocomm(comm, taxo, rank)
}
\arguments{
\item{comm}{community format data.frame (samples in rows, taxa in columns, rows and columns have informative names)}

\item{taxo}{a data frame with rows in same order as columns of \code{comm} and taxonomic information for each row provided in columns}

\item{rank}{Taxonomic rank to aggregate at (a column name in \code{taxo})}
}
\value{
A community data.frame where columns are taxonomic groups
}
\description{
Create taxonomy-aggregated community data.frame
}
