% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_expression.R
\name{upperQuartile}
\alias{upperQuartile}
\title{Upper Quartile normalize expression matrix}
\usage{
upperQuartile(expression_matrix)
}
\arguments{
\item{log2}{expression matrix}
}
\value{
log2 quartile normalized expression
}
\description{
Upper Quartile normalize expression matrix
}
\examples{
upperQuartile(expression_matrix, reference_matrix)
}
\keyword{expression}

