% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{s2c}
\alias{c2s}
\alias{base_complement}
\alias{reverse_complement}
\title{Useful Utilitys for Genomic Data}
\usage{
s2c(string)

c2s(chr_vector)

base_complement(base)

reverse_complement(Sequence)
}
\description{
Useful Utilitys for Genomic Data
}
