% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_helpers.R
\name{write.df}
\alias{write.df}
\title{simplfy options for writing data frame}
\usage{
write.df(input_df, file = NULL)
}
\arguments{
\item{data}{frame}

\item{output}{file (optional)}
}
\description{
simplfy options for writing data frame
}
