% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.fasta.R
\name{write.fasta}
\alias{write.fasta}
\title{Write List(SeqFasta) to Fasta Format}
\usage{
write.fasta(lst, fname)
}
\arguments{
\item{fname}{file name to write to}

\item{list}{list of 'SeqFasta' objects}
}
\description{
Write List(SeqFasta) to Fasta Format
}
\examples{
write.fasta(list(SeqFasta), 'output.fa')

}
