% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.fastq.R
\name{write.fastq}
\alias{write.fastq}
\title{Write List(SeqFastq) to Fastq Format}
\usage{
write.fastq(lst, fname)
}
\arguments{
\item{fname}{file name to write to}

\item{list}{list of 'SeqFastq' objects}
}
\description{
Write List(SeqFastq) to Fastq Format
}
\examples{
write.fastq(list(SeqFastq), 'output.fq')

}
