% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.q.R
\name{avg.q}
\alias{avg.q}
\alias{avg.q.SeqFastq}
\title{Calculate Average Qscore}
\usage{
avg.q(object)

\method{avg.q}{SeqFastq}(object)
}
\arguments{
\item{object}{class 'SeqFastq' object}
}
\description{
Generic function calculates average qscore of a S3 'SeqFastq' object.
}
\examples{
avg.q()
}
