% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr_transform}
\alias{clr_transform}
\title{Centered log-ratio (CLR) transform community data}
\usage{
clr_transform(x)
}
\arguments{
\item{x}{A vegan-format matrix of samples by taxa}
}
\value{
CLR-transformed matrix
}
\description{
Centered log-ratio (CLR) transform community data
}
