% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarizePhenotype.R
\name{summary.phenotype}
\alias{summary.phenotype}
\title{Summarize the phenotype data frame}
\usage{
\method{summary}{phenotype}(p, max.levels = 5)
}
\arguments{
\item{p}{an object of class 'phenotype'.}

\item{max.levels}{maximum number of \code{levels}.  See Details}
}
\value{
an object of class 'phenotype'.  \code{data} is minimal, \code{include} is set to, \code{exclude} is set to
}
\description{
This function summarizes the phenotype data in a phenotype object.
}
\details{
This function calcualtes basic summary statistics for phenotype data. [TBD: table of summary statsitcs based on data type].

If the number of unique character, integer or factor levels is less than or equal to max.levels
}

